+++ src/traces/image/attributes.js
@@ -8,7 +8,7 @@
 
 'use strict';
 
-var plotAttrs = require('../../plots/attributes');
+var baseAttrs = require('../../plots/attributes');
 var hovertemplateAttrs = require('../../plots/template_attributes').hovertemplateAttrs;
 var extendFlat = require('../../lib/extend').extendFlat;
 var colormodel = require('./constants').colormodel;
@@ -35,19 +35,19 @@ module.exports = extendFlat({
         values: cm,
         dflt: 'rgb',
         role: 'info',
-        editType: 'plot',
+        editType: 'calc',
         description: 'Color model used to map the numerical color components described in `z` into colors.'
     },
     zmin: {
         valType: 'info_array',
         items: [
-            {valType: 'number', editType: 'plot'},
-            {valType: 'number', editType: 'plot'},
-            {valType: 'number', editType: 'plot'},
-            {valType: 'number', editType: 'plot'}
+            {valType: 'number', editType: 'calc'},
+            {valType: 'number', editType: 'calc'},
+            {valType: 'number', editType: 'calc'},
+            {valType: 'number', editType: 'calc'}
         ],
         role: 'info',
-        editType: 'plot',
+        editType: 'calc',
         description: [
             'Array defining the lower bound for each color component.',
             'Note that the default value will depend on the colormodel.',
@@ -57,13 +57,13 @@ module.exports = extendFlat({
     zmax: {
         valType: 'info_array',
         items: [
-            {valType: 'number', editType: 'plot'},
-            {valType: 'number', editType: 'plot'},
-            {valType: 'number', editType: 'plot'},
-            {valType: 'number', editType: 'plot'}
+            {valType: 'number', editType: 'calc'},
+            {valType: 'number', editType: 'calc'},
+            {valType: 'number', editType: 'calc'},
+            {valType: 'number', editType: 'calc'}
         ],
         role: 'info',
-        editType: 'plot',
+        editType: 'calc',
         description: [
             'Array defining the higher bound for each color component.',
             'Note that the default value will depend on the colormodel.',
@@ -108,7 +108,7 @@ module.exports = extendFlat({
         editType: 'plot',
         description: 'Same as `text`.'
     },
-    hoverinfo: extendFlat({}, plotAttrs.hoverinfo, {
+    hoverinfo: extendFlat({}, baseAttrs.hoverinfo, {
         flags: ['x', 'y', 'z', 'color', 'name', 'text'],
         dflt: 'x+y+z+text+name'
     }),